# frontend/openapi.yaml
openapi: 3.0.0
info:
  title: Stark Byte Hub API
  version: 1.0.0
  description: Documentation de l’API du backend Stark Byte Hub
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
paths:
...
  /download/firmwares/{filename}:
    get:
      summary: Télécharge un fichier firmware
      tags:
        - Download
      security:
        - cookieAuth: []
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: Nom du fichier firmware
      responses:
        "200":
          description: Fichier téléchargé
        "404":
          description: Fichier introuvable
...
  /news:
    get:
      summary: Liste toutes les actualités
      tags:
        - News
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Liste des actualités
        "403":
          description: Permission refusée
    post:
      summary: Crée une nouvelle actualité
      tags:
        - News
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
              properties:
                title:
                  type: string
                  example: Nouvelle mise à jour disponible
                description:
                  type: string
                  example: Une nouvelle version du firmware est disponible...
                tags:
                  type: array
                  items:
                    type: string
                  example:
                    - mise à jour
                    - firmware
                date:
                  type: string
                  format: date
                  example: 2025-04-04
                images:
                  type: array
                  items:
                    type: string
                  example:
                    - /uploads/news/image1.png
                coverImageId:
                  type: string
                  nullable: true
      responses:
        "201":
          description: Actualité créée
        "400":
          description: Données invalides
        "403":
          description: Permission refusée
  /news/{id}:
    get:
      summary: Détail d’une actualité
      tags:
        - News
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: ID de l’actualité
          schema:
            type: string
      responses:
        "200":
          description: Actualité trouvée
        "403":
          description: Permission refusée
        "404":
          description: Actualité non trouvée
    patch:
      summary: Met à jour une actualité existante
      tags:
        - News
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: ID de l’actualité
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                date:
                  type: string
                  format: date
                images:
                  type: array
                  items:
                    type: string
                coverImageId:
                  type: string
                  nullable: true
      responses:
        "200":
          description: Actualité mise à jour
        "403":
          description: Permission refusée
        "404":
          description: Actualité non trouvée
    delete:
      summary: Supprime une actualité
      tags:
        - News
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: ID de l’actualité
          schema:
            type: string
      responses:
        "204":
          description: Actualité supprimée
        "403":
          description: Permission refusée
        "404":
          description: Actualité non trouvée
  /permissions:
    get:
      summary: Liste toutes les permissions
      tags:
        - Permissions
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Liste des permissions
        "403":
          description: Permission refusée
    post:
      summary: Crée une nouvelle permission
      tags:
        - Permissions
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Nom unique de la permission
      responses:
        "201":
          description: Permission créée avec succès
        "400":
          description: Données invalides
        "403":
          description: Permission refusée
  /permissions/{id}:
    get:
      summary: Détail d’une permission
      tags:
        - Permissions
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la permission
          schema:
            type: string
      responses:
        "200":
          description: Permission trouvée
        "403":
          description: Permission refusée
        "404":
          description: Permission introuvable
    patch:
      summary: Met à jour une permission
      tags:
        - Permissions
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la permission
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nouveau nom de la permission
      responses:
        "200":
          description: Permission mise à jour
        "400":
          description: Données invalides
        "403":
          description: Permission refusée
        "404":
          description: Permission introuvable
    delete:
      summary: Supprime une permission
      tags:
        - Permissions
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la permission
          schema:
            type: string
      responses:
        "204":
          description: Permission supprimée
        "403":
          description: Permission refusée
        "404":
          description: Permission introuvable
...